chore: renamed supported locale file
